refactor timeline store for code splitting
