+++ web/ce/store/timeline/index.ts
@@ -1,0 +1,23 @@
+import { RootStore } from "@/plane-web/store/root.store";
+import { IIssuesTimeLineStore, IssuesTimeLineStore } from "@/store/timeline/issues-timeline.store";
+import { IModulesTimeLineStore, ModulesTimeLineStore } from "@/store/timeline/modules-timeline.store";
+import { BaseTimeLineStore, IBaseTimelineStore } from "./base-timeline.store";
+
+export interface ITimelineStore {
+  issuesTimeLineStore: IIssuesTimeLineStore;
+  modulesTimeLineStore: IModulesTimeLineStore;
+  projectTimeLineStore: IBaseTimelineStore;
+}
+
+export class TimeLineStore implements ITimelineStore {
+  issuesTimeLineStore: IIssuesTimeLineStore;
+  modulesTimeLineStore: IModulesTimeLineStore;
+  projectTimeLineStore: IBaseTimelineStore;
+
+  constructor(rootStore: RootStore) {
+    this.issuesTimeLineStore = new IssuesTimeLineStore(rootStore);
+    this.modulesTimeLineStore = new ModulesTimeLineStore(rootStore);
+    // Dummy store
+    this.projectTimeLineStore = new BaseTimeLineStore(rootStore);
+  }
+}
