+++ web/core/components/gantt-chart/contexts/index.tsx
@@ -3,6 +3,7 @@ import { createContext, useContext } from "react";
 export enum ETimeLineTypeType {
   ISSUE = "ISSUE",
   MODULE = "MODULE",
+  PROJECT = "PROJECT",
 }
 
 export const TimeLineTypeContext = createContext<ETimeLineTypeType | undefined>(undefined);
