+++ web/core/components/gantt-chart/sidebar/utils.ts
@@ -16,17 +16,17 @@ export const handleOrderChange = (
   // return if dropped outside the list
   if (sourceBlockIndex === -1 || destinationBlockIndex === -1 || sourceBlockIndex === destinationBlockIndex) return;
 
-  let updatedSortOrder = getBlockById(blockIds[sourceBlockIndex])?.sort_order;
+  let updatedSortOrder = getBlockById(blockIds[sourceBlockIndex])?.sort_order ?? 0;
 
   // update the sort order to the lowest if dropped at the top
-  if (destinationBlockIndex === 0) updatedSortOrder = getBlockById(blockIds[0])?.sort_order - 1000;
+  if (destinationBlockIndex === 0) updatedSortOrder = (getBlockById(blockIds[0])?.sort_order ?? 0) - 1000;
   // update the sort order to the highest if dropped at the bottom
   else if (destinationBlockIndex === blockIds.length)
-    updatedSortOrder = getBlockById(blockIds[blockIds.length - 1])?.sort_order + 1000;
+    updatedSortOrder = (getBlockById(blockIds[blockIds.length - 1])?.sort_order ?? 0) + 1000;
   // update the sort order to the average of the two adjacent blocks if dropped in between
   else {
-    const destinationSortingOrder = getBlockById(blockIds[destinationBlockIndex])?.sort_order;
-    const relativeDestinationSortingOrder = getBlockById(blockIds[destinationBlockIndex - 1])?.sort_order;
+    const destinationSortingOrder = getBlockById(blockIds[destinationBlockIndex])?.sort_order ?? 0;
+    const relativeDestinationSortingOrder = getBlockById(blockIds[destinationBlockIndex - 1])?.sort_order ?? 0;
 
     updatedSortOrder = (destinationSortingOrder + relativeDestinationSortingOrder) / 2;
   }
