+++ web/core/store/timeline/issues-timeline.store.ts
@@ -8,13 +8,10 @@ export interface IIssuesTimeLineStore extends IBaseTimelineStore {
 }
 
 export class IssuesTimeLineStore extends BaseTimeLineStore implements IIssuesTimeLineStore {
-  isDependencyEnabled = true;
-
   constructor(_rootStore: RootStore) {
     super(_rootStore);
 
-    autorun((reaction) => {
-      reaction.trace();
+    autorun(() => {
       const getIssueById = this.rootStore.issue.issues.getIssueById;
       this.updateBlocks(getIssueById);
     });
