+++ web/core/store/timeline/modules-timeline.store.ts
@@ -8,13 +8,10 @@ export interface IModulesTimeLineStore extends IBaseTimelineStore {
 }
 
 export class ModulesTimeLineStore extends BaseTimeLineStore implements IModulesTimeLineStore {
-  isDependencyEnabled = false;
-
   constructor(_rootStore: RootStore) {
     super(_rootStore);
 
-    autorun((reaction) => {
-      reaction.trace();
+    autorun(() => {
       const getModuleById = this.rootStore.module.getModuleById;
       this.updateBlocks(getModuleById);
     });
